Clean up DMath.kt

+++ dsensor/src/main/java/com/hoan/dsensor/utils/DMath.kt
@@ -8,7 +8,7 @@ import kotlin.math.*
  */
 
 fun calculateNorm(vector: FloatArray): Float {
-    var norm = 0.toFloat()
+    var norm = 0f
     for (coordinate in vector) {
         norm += coordinate.pow(2)
     }
@@ -16,21 +16,20 @@ fun calculateNorm(vector: FloatArray): Float {
 }
 
 fun addAngle(angle: Float, currentSum: FloatArray) {
-    currentSum[0] += sin(angle.toDouble()).toFloat()
-    currentSum[1] += cos(angle.toDouble()).toFloat()
+    currentSum[0] += sin(angle)
+    currentSum[1] += cos(angle)
 }
 
 fun removeAngle(angle: Float, currentSum: FloatArray) {
-    currentSum[0] -= sin(angle.toDouble()).toFloat()
-    currentSum[1] -= cos(angle.toDouble()).toFloat()
+    currentSum[0] -= sin(angle)
+    currentSum[1] -= cos(angle)
 }
 
 fun averageAngle(currentSum: FloatArray, totalTerms: Int): Float {
-    return atan2((currentSum[0] / totalTerms).toDouble(), (currentSum[1] / totalTerms).toDouble()).toFloat()
+    return atan2(currentSum[0] / totalTerms, currentSum[1] / totalTerms)
 }
 
 fun scaleVector(vector: FloatArray, scaleFactor: Float): FloatArray {
-    //logger("DMathKt", "scaleVector(" + vector[0] + ", " + vector[1] + ", " + vector[2] + ", " + scaleFactor + ")")
     val result = FloatArray(vector.size)
     for (i in vector.indices) {
         result[i] = scaleFactor * vector[i]
@@ -39,8 +38,6 @@ fun scaleVector(vector: FloatArray, scaleFactor: Float): FloatArray {
 }
 
 fun productOfSquareMatrixAndVector(matrix: FloatArray?, vector: FloatArray?): FloatArray? {
-    //logger("DMathKt", "productOfSquareMatrixAndVector: vector = ("
-    //        + vector[0] + ", " + vector[1] + ", " + vector[2] + ")");
     if (matrix == null || vector == null || matrix.size != vector.size * vector.size) {
         return null
     }
